Fix handling login attempt with unconfirmed account
