+++ test/dealog_backoffice_web/controllers/user_session_controller_test.exs
@@ -6,7 +6,9 @@ defmodule DealogBackofficeWeb.UserSessionControllerTest do
   import DealogBackoffice.AccountsFixtures
 
   setup do
-    %{user: user_fixture()}
+    user = user_fixture()
+    confirm_user(user)
+    %{user: user}
   end
 
   describe "GET /users/log_in" do
