+++ test/support/conn_case.ex
@@ -24,6 +24,7 @@ defmodule DealogBackofficeWeb.ConnCase do
       import Phoenix.ConnTest
       import DealogBackofficeWeb.ConnCase
       import DealogBackofficeWeb.LiveViewTestHelpers
+      import DealogBackoffice.AccountTestHelpers
 
       alias DealogBackofficeWeb.Router.Helpers, as: Routes
 
@@ -40,58 +41,5 @@ defmodule DealogBackofficeWeb.ConnCase do
     end
 
     {:ok, conn: Phoenix.ConnTest.build_conn()}
-  end
-
-  @doc """
-  Setup helper that registers users.
-
-      setup :register_user
-
-  It stores an updated connection in the test context.
-  """
-  def register_user(%{conn: conn}) do
-    user = DealogBackoffice.AccountsFixtures.user_fixture()
-
-    %{conn: log_in_user(conn, user), user: user}
-  end
-
-  @doc """
-  Setup helper that registers, confirms and logs in users.
-
-      setup :register_and_log_in_user
-
-  It stores an updated connection and a registered user in the
-  test context.
-  """
-  def register_and_log_in_user(%{conn: conn}) do
-    user = DealogBackoffice.AccountsFixtures.user_fixture()
-    confirm_user(user)
-
-    %{conn: log_in_user(conn, user), user: user}
-  end
-
-  @doc """
-  Confirm user registration. The user is only able to login / create a session
-  once they confirmed their registration.
-  """
-  def confirm_user(user) do
-    {encoded_token, user_token} =
-      DealogBackoffice.Accounts.UserToken.build_email_token(user, "confirm")
-
-    DealogBackoffice.Repo.insert!(user_token)
-    DealogBackoffice.Accounts.confirm_user(encoded_token)
-  end
-
-  @doc """
-  Logs the given `user` into the `conn`.
-
-  It returns an updated `conn`.
-  """
-  def log_in_user(conn, user) do
-    token = DealogBackoffice.Accounts.generate_user_session_token(user)
-
-    conn
-    |> Phoenix.ConnTest.init_test_session(%{})
-    |> Plug.Conn.put_session(:user_token, token)
   end
 end
